backend/requirements.txt
# Construction Photo Analyzer - Backend Dependencies
# Generated with GPU support (CUDA 12.4)

# ============================================
# Core Deep Learning
# ============================================
# PyTorch with CUDA support - install separately:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0

# ============================================
# Web Framework
# ============================================
fastapi>=0.127.0
uvicorn[standard]>=0.40.0
python-multipart>=0.0.21

# ============================================
# Data Validation & Configuration
# ============================================
pydantic>=2.12.0
pydantic-settings>=2.0.0
python-dotenv>=1.2.0

# ============================================
# Image Processing
# ============================================
pillow>=12.0.0
numpy>=2.3.0

# ============================================
# HTTP Client
# ============================================
httpx>=0.28.0

# ============================================
# Development & Testing
# ============================================
pytest>=9.0.0
pytest-asyncio>=1.3.0
pytest-cov>=7.0.0
black>=25.0.0
flake8>=7.3.0
